refactor(stubs): utilize locator strategies over stub source factory method.

+++ micropy/stubs/source.py
@@ -14,91 +14,72 @@ import tempfile
 from contextlib import contextmanager
 from functools import partial
 from pathlib import Path
-from typing import Any, Callable, Optional, final
+from typing import Any, Callable, ContextManager, Iterable, Optional, Union
 
+import attrs
 from micropy import utils
 from micropy.logger import Log
 from micropy.utils.types import PathStr
+from typing_extensions import Protocol
 
 
-class StubSource(abc.ABC):
-    """Abstract Base Class for Stub Sources."""
-
-    location: PathStr
-
-    def __init__(self, location: PathStr, log=None):
-        self.location = location
-        _name = self.__class__.__name__
-        self.log = log or Log.add_logger(_name)
-
+class LocateStrategy(Protocol):
     @abc.abstractmethod
-    def prepare(self) -> tuple[PathStr, Optional[Callable[..., Any]]]:
-        """Prepares the source."""
+    def prepare(self, location: PathStr) -> Union[PathStr, tuple[PathStr, Callable[..., Any]]]:
+        ...
+
+
+logger = Log.add_logger(__name__)
+
+
+@attrs.define
+class StubSource:
+    """Handles sourcing stubs."""
+
+    location: PathStr = attrs.field()
+    locators: list[LocateStrategy] = attrs.field()
+
+    @locators.default
+    def _default_locators(self: StubSource) -> list[LocateStrategy]:
+        return [RemoteStubLocator(), StubInfoSpecLocator()]
 
     @contextmanager
-    def ready(self):
+    def ready(self) -> ContextManager[PathStr]:
         """Yields prepared Stub Source.
 
         Allows StubSource subclasses to have a preparation
         method before providing a local path to itself.
 
-        Args:
-            path (str, optional): path to stub source.
-                Defaults to location.
-            teardown (func, optional): callback to execute on exit.
-                Defaults to None.
-
         Yields:
             Resolved PathLike object to stub source
 
         """
-        prep = iter(self.prepare())
-        path = next(prep, self.location)
-        teardown = next(prep, lambda: None)
-        info_path = next(Path(path).rglob("info.json"), None)
-        path = Path(info_path.parent) if info_path else path
+        path = self.location
+        teardown = lambda: None
+        for locator in self.locators:
+            logger.debug(f"running (strategy:{locator}) @ (location:{path})")
+            response = locator.prepare(path)
+            parts = iter(response if isinstance(response, Iterable) else (response,))
+            path = next(parts, path)
+            teardown = next(parts, teardown)
+            logger.debug(f"results of (strategy:{locator}) -> (location:{path})")
         yield path
         if teardown:
             teardown()
 
-    def __str__(self):
-        _name = self.__class__.__name__
-        return f"<{_name}@{self.location}>"
+
+@attrs.define
+class StubInfoSpecLocator(LocateStrategy):
+    def prepare(self, location: PathStr) -> PathStr:
+        info_path = next(Path(location).rglob("info.json"), None)
+        return location if info_path is None else info_path.parent
 
 
-@final
-class LocalStubSource(StubSource):
-    """Stub Source Subclass for local locations.
+@attrs.define
+class RemoteStubLocator(LocateStrategy):
+    """Stub Source for remote locations."""
 
-    Args:
-        path (str): Path to Stub Source
-
-    Returns:
-        obj: Instance of LocalStubSource
-
-    """
-
-    def prepare(self) -> tuple[PathStr, Optional[Callable[..., Any]]]:
-        return self.location, None
-
-    def __init__(self, path, **kwargs):
-        location = utils.ensure_existing_dir(path)
-        super().__init__(location, **kwargs)
-
-
-@final
-class RemoteStubSource(StubSource):
-    """Stub Source for remote locations.
-
-    Args:
-        url (str): URL to Stub Source
-
-    Returns:
-        obj: Instance of RemoteStubSource
-
-    """
-
-    def _unpack_archive(self, file_bytes, path):
+    def _unpack_archive(self, file_bytes: bytes, path: PathStr) -> PathStr:
         """Unpack archive from bytes buffer.
 
         Args:
@@ -114,25 +95,22 @@ class RemoteStubSource(StubSource):
         output = next(path.iterdir())
         return output
 
-    def prepare(self) -> tuple[PathStr, Optional[Callable[..., Any]]]:
+    def prepare(self, location: PathStr) -> tuple[PathStr, Optional[Callable[..., Any]]] | PathStr:
         """Retrieves and unpacks source.
 
         Prepares remote stub resource by downloading and
         unpacking it into a temporary directory.
-        This directory is removed on exit of the superclass
-        context manager
-
-        Returns:
-            callable: StubSource.ready parent method
+        This directory is removed via the returned teardown.
 
         """
+        if not utils.is_url(location):
+            logger.debug(f"{self}: {location} not viable, skipping...")
+            return location
         tmp_dir = tempfile.mkdtemp()
         tmp_path = Path(tmp_dir)
-        filename = utils.get_url_filename(self.location).split(".tar.gz")[0]
-        _file_name = "".join(self.log.iter_formatted(f"$B[{filename}]"))
-        content = utils.stream_download(
-            self.location, desc=f"{self.log.get_service()} {_file_name}"
-        )
+        filename = utils.get_url_filename(location).split(".tar.gz")[0]
+        _file_name = "".join(logger.iter_formatted(f"$B[{filename}]"))
+        content = utils.stream_download(location, desc=f"{logger.get_service()} {_file_name}")
         source_path = self._unpack_archive(content, tmp_path)
         teardown = partial(shutil.rmtree, tmp_path)
         return source_path, teardown
@@ -141,6 +119,8 @@ class RemoteStubSource(StubSource):
 def get_source(location, **kwargs):
     """Factory for StubSource Instance.
 
+    Deprecated. Todo: Remove.
+
     Args:
         location (str): PathLike object or valid URL
 
@@ -148,9 +128,4 @@ def get_source(location, **kwargs):
         obj: Either Local or Remote StubSource Instance
 
     """
-    try:
-        utils.ensure_existing_dir(location)
-    except NotADirectoryError:
-        return RemoteStubSource(location, **kwargs)
-    else:
-        return LocalStubSource(location, **kwargs)
+    return StubSource(location, **kwargs)
